backend/brevio_service/requirements.txt
python-dotenv
openai-whisper==20240930
openai==1.57.2
ffmpeg